+++ custom/Extension/application/Ext/Extensions/PrvCloud.php
@@ -1,0 +1,6 @@
+<?php
+$extensions["prv_cloud"] =  array(
+    "section" => "prv_cloud_class",
+    "extdir" => "PrvCloud",
+    "file" => 'prv_cloud.ext.php',
+    "module" => "");
